generic conv working, added 1 test, removed some warnings, todo tests vs reference, tests vs caffe

+++ tests/test_cases/relu_test.cpp
@@ -82,7 +82,7 @@ TEST(relu_f32_fw, intrinsics_avx2) {
     execute({output, opt_relu}).sync();
     execute({ref_output, ref_relu}).sync();
 
-    for(int output_element = 0; output_element < output_memory.count(); ++output_element)
+    for(size_t output_element = 0; output_element < output_memory.count(); ++output_element)
         EXPECT_EQ(true, tests::are_equal(get_value<float>(ref_output_memory, output_element), get_value<float>(output_memory, output_element)));
 }
 
@@ -209,7 +209,7 @@ TEST(relu_f32_bw, intrinsics_avx2) {
     execute({bw_output, opt_relu}).sync();
     execute({ref_bw_output, ref_relu}).sync();
 
-    for(int output_element = 0; output_element < bw_output_memory.count(); ++output_element)
+    for(size_t output_element = 0; output_element < bw_output_memory.count(); ++output_element)
         EXPECT_EQ(true, tests::are_equal(get_value<float>(ref_bw_output_memory, output_element), get_value<float>(bw_output_memory, output_element)));
 }
